Add a g:tslime_ensure_trailing_newlines option.
This is handy when you want to evaluate part of a line as a line of its own.

If this number is non-zero it will ensure that at least that many trailing
newlines are sent every time.  This is useful for Python where you might want to
set it to "2" so indented functions/classes are easier to send.